refactor member mutation
