+++ libs/data-layer/operations/mutations/index.ts
@@ -1,6 +1,6 @@
 export * from './useCreateFamilyMutation';
+export * from './useCreateMemberMutation';
 export * from './useCreateProjectMutation';
 export * from './useDeleteFamilyMutation';
 export * from './useLoginMutation';
-export * from './useMemberMutation';
 export * from './useUpsertHouseholderMutation';
